WIP refactoring tests to share more code
