+++ test/process_cache_test.exs
@@ -3,13 +3,7 @@ defmodule ProcessCacheTest do
 
   alias PayDayLoan, as: PDL
 
-  # repeatedly call pred_callback until it returns true
-  #   - no dwell between iterations and no timeout
-  def wait_for(pred_callback) do
-    Stream.repeatedly(fn -> !pred_callback.() end)
-    |> Stream.drop_while(&(&1))
-    |> Enum.take(1)
-  end
+  import PayDayLoanTest.Support, only: [wait_for: 1]
 
   alias PayDayLoanTest.Support.LoadHistory
 
@@ -27,150 +21,54 @@ defmodule ProcessCacheTest do
 
   # loader behaviour implementation
   defmodule PDLTestProcessImplementation do
-    @behaviour PayDayLoan.Loader
+    use PayDayLoan.Support.TestImplementation
 
-    @key_that_shall_be_replaced "key that shall be replaced" 
-    @key_that_does_not_exist "key that does not exist"
-    @key_that_will_not_new "key that will not new"
-    @key_that_will_not_refresh "key that will not refresh"
-    @key_that_returns_ignore_on_new "key that returns ignore on new"
-    @key_that_returns_ignore_on_refresh "key that returns ignore on refresh"
-
-    # we'll refuse to load this key
-    def key_that_shall_not_be_loaded do
-      "key that shall not be loaded"
-    end
-
-    # this one will get a new pid
-    def key_that_shall_be_replaced do
-      @key_that_shall_be_replaced
-    end
-
-    # this one will fail the key cache check
-    def key_that_does_not_exist do
-      @key_that_does_not_exist
-    end
-
-    # this key will return an error on new
-    def key_that_will_not_new do
-      @key_that_will_not_new
-    end
-
-    # this key will return an error on refresh
-    def key_that_will_not_refresh do
-      @key_that_will_not_refresh
-    end
-
-    # this key will return :ignore on new
-    def key_that_returns_ignore_on_new do
-      @key_that_returns_ignore_on_new
-    end
-
-    def key_that_returns_ignore_on_refresh do
-      @key_that_returns_ignore_on_refresh
-    end
-
-    def bulk_load(keys) do
-      LoadHistory.loaded(keys)
-
-      keys = keys
-      |> Enum.filter(fn(k) ->
-        k != key_that_shall_not_be_loaded()
-      end)
-
-      Enum.map(keys, fn(key) -> {key, key} end)
-    end
-
-    def new(key = @key_that_will_not_new, value) do
-      LoadHistory.new(key, value)
-      {:error, :load_failed}
-    end
-    def new(key = @key_that_returns_ignore_on_new, value) do
-      LoadHistory.new(key, value)
-      :ignore
-    end
-    def new(key, value) do
-      LoadHistory.new(key, value)
+    def on_new(key, value) do
       Agent.start(fn() -> [{key, value}] end)
     end
 
-    def refresh(pid, key = @key_that_will_not_refresh, value) do
-      LoadHistory.refresh(pid, key, value)
-      # stop the previous pid
-      Agent.stop(pid)
-      # return error
-      {:error, :refresh_failed}
-    end
-    def refresh(pid, key = @key_that_returns_ignore_on_refresh, value) do
-      LoadHistory.refresh(pid, key, value)
-      # stop the previous pid
-      Agent.stop(pid)
-      # return ignore
-      :ignore
-    end
-    def refresh(pid, key = @key_that_shall_be_replaced, value) do
-      LoadHistory.refresh(pid, key, value)
-      # stop the previous pid
-      Agent.stop(pid)
-      # create a new one
-      Agent.start(fn() -> [{key, value}] end)
-    end
-    def refresh(pid, key, value) do
-      LoadHistory.refresh(pid, key, value)
-      Agent.update(pid, fn(values) -> [{key, value} | values] end)
-      {:ok, pid}
+    def on_remove(_key, value) do
+      Agent.stop(value)
     end
 
-    def key_exists?(@key_that_does_not_exist), do: false
-    def key_exists?(_key), do: true
+    def on_replace(old_value, key, value) do
+      on_remove(key, old_value)
+      on_new(key, value)
+    end
+
+    def on_update(old_value, key, value) do
+      Agent.update(old_value, fn(values) -> [{key, value} | values] end)
+      {:ok, old_value}
+    end
   end
 
-  setup do
-    LoadHistory.start
-
-    wait_for(fn -> Process.whereis(PDLTestSup) == nil end)
-
-    pdl = PDLTestProcessCache.pdl
-    sup_spec = PayDayLoan.supervisor_specification(pdl)
-
-    {:ok, _sup_pid} = Supervisor.start_link(
-      [sup_spec],
-      strategy: :one_for_one,
-      name: PDLTestSup
-    )
-
-    on_exit fn ->
-      LoadHistory.stop
-    end
-
-    :ok
+  def pdl do
+    PDLTestProcessCache.pdl()
   end
+
+  use PayDayLoan.Support.CommonTests
 
   test "basic integration test" do
-    pdl = PDLTestProcessCache.pdl
+    assert 0 == PDL.EtsBackend.size(pdl())
 
-    assert 0 == PDL.EtsBackend.size(pdl)
+    assert {:error, :not_found} == PDL.peek_pid(pdl(), 1)
 
-    assert {:error, :not_found} == PDL.peek_pid(pdl, 1)
+    assert :requested == PDL.query_load_state(pdl(), 1)
 
-    assert :requested == PDL.query_load_state(pdl, 1)
-
-    GenServer.cast(pdl.load_worker, :ping)
+    GenServer.cast(pdl().load_worker, :ping)
 
     {:ok, pid} = PDLTestProcessCache.get_pid(1)
 
-    assert :loaded == PDL.query_load_state(pdl, 1)
+    assert :loaded == PDL.query_load_state(pdl(), 1)
 
-    assert {:ok, pid} == PDL.peek_pid(pdl, 1)
+    assert {:ok, pid} == PDL.peek_pid(pdl(), 1)
 
     assert [{1, 1}] == Agent.get(pid, fn(v) -> v end)
 
-    assert PDL.KeyCache.in_cache?(pdl.key_cache, 1)
+    assert PDL.KeyCache.in_cache?(pdl().key_cache, 1)
   end
 
   test "loading happens in bulk" do
-    pdl = PDLTestProcessCache.pdl
-
     n = 10
 
     tasks = (1..n)
@@ -189,12 +87,11 @@ defmodule ProcessCacheTest do
 
     (1..n)
     |> Enum.each(fn(ix) ->
-      assert PDL.KeyCache.in_cache?(pdl.key_cache, ix)
+      assert PDL.KeyCache.in_cache?(pdl().key_cache, ix)
     end)
   end
 
   test "refreshing a cache element" do
-    pdl = PDLTestProcessCache.pdl
     replaced_key = PDLTestProcessImplementation.key_that_shall_be_replaced
 
     {:ok, pid1} = PDLTestProcessCache.get_pid(1)
@@ -212,7 +109,7 @@ defmodule ProcessCacheTest do
 
     wait_for(
       fn ->
-        PDL.LoadState.any_requested?(pdl.load_state_manager)
+        PDL.LoadState.any_requested?(pdl().load_state_manager)
       end
     )
 
@@ -233,34 +130,29 @@ defmodule ProcessCacheTest do
   end
 
   test "requesting a key that is in key cache but fails to load" do
-    # reduce the load wait time to make the test go by faster
-    pdl = %{ PDLTestProcessCache.pdl | load_wait_msec: 10 }
-
     key = PDLTestProcessImplementation.key_that_shall_not_be_loaded
 
     assert {:error, :failed} == PDLTestProcessCache.get_pid(key)
     assert [{:loaded, [key]}] == LoadHistory.loads
-    assert nil == PDL.LoadState.peek(pdl.load_state_manager, key)
+    assert nil == PDL.LoadState.peek(pdl().load_state_manager, key)
     # we hold onto the knowledge that the key exists
-    assert PDL.KeyCache.in_cache?(pdl.key_cache, key)
+    assert PDL.KeyCache.in_cache?(pdl().key_cache, key)
   end
 
   test "requesting a key that does not exist" do
-    pdl = PDLTestProcessCache.pdl
-
     key = PDLTestProcessImplementation.key_that_does_not_exist
 
     assert {:error, :not_found} == PDLTestProcessCache.get_pid(key)
     assert [] == LoadHistory.loads
-    refute PDL.KeyCache.in_cache?(pdl.key_cache, key)
+    refute PDL.KeyCache.in_cache?(pdl().key_cache, key)
   end
 
   test "overriding defaults" do
-    assert 10 == PDLTestProcessCache.pdl.batch_size
+    assert 10 == pdl().batch_size
   end
 
   test "large batches don't require an extra ping" do
-    batch_size = PDLTestProcessCache.pdl.batch_size
+    batch_size = pdl().batch_size
     keys = Enum.to_list(1..(2 * batch_size))
 
     PDLTestProcessCache.request_load(keys)
@@ -274,7 +166,7 @@ defmodule ProcessCacheTest do
     key = PDLTestProcessImplementation.key_that_will_not_new
     assert {:error, :failed} == PDLTestProcessCache.get_pid(key)
     # should get cleared from the load state cache
-    assert nil == PDL.peek_load_state(PDLTestProcessCache.pdl, key)
+    assert nil == PDL.peek_load_state(pdl(), key)
   end
 
   test "refresh failures are ignored (should be handled in callback)" do
@@ -291,7 +183,7 @@ defmodule ProcessCacheTest do
     # fail to refresh
     assert {:error, :failed} == PDLTestProcessCache.get_pid(key)
     # should get cleared from the load state cache
-    assert nil == PDL.peek_load_state(PDLTestProcessCache.pdl, key)
+    assert nil == PDL.peek_load_state(pdl(), key)
   end
 
   test "working with key/pid lists" do
@@ -335,7 +227,7 @@ defmodule ProcessCacheTest do
 
     Agent.stop(pid1)
 
-    assert {:error, :not_found} == PDL.peek_pid(PDLTestProcessCache.pdl, 1)
+    assert {:error, :not_found} == PDL.peek_pid(pdl(), 1)
 
     {:ok, pid2} = PDLTestProcessCache.get_pid(1)
 
@@ -344,21 +236,19 @@ defmodule ProcessCacheTest do
   end
 
   test "when the monitor is killed, it restarts and starts monitoring again" do
-    pdl = PDLTestProcessCache.pdl
-
     {:ok, pid1} = PDLTestProcessCache.get_pid(1)
 
-    previous_pid = Process.whereis(pdl.cache_monitor)
+    previous_pid = Process.whereis(pdl().cache_monitor)
 
-    :ok = GenServer.stop(pdl.cache_monitor)
+    :ok = GenServer.stop(pdl().cache_monitor)
 
     # make sure that it restarts via the supervisor
     wait_for(fn ->
-      pid = Process.whereis(pdl.cache_monitor)
+      pid = Process.whereis(pdl().cache_monitor)
       is_pid(pid) && Process.alive?(pid)
     end)
 
-    refute previous_pid == Process.whereis(pdl.cache_monitor)
+    refute previous_pid == Process.whereis(pdl().cache_monitor)
 
     # the new monitor should be monitoring the pid now
     Agent.stop(pid1)
@@ -385,7 +275,7 @@ defmodule ProcessCacheTest do
 
     assert {:ok, pid} == PDLTestProcessCache.get_pid(1)
 
-    :ok = PayDayLoan.uncache_key(PDLTestProcessCache.pdl, 1)
+    :ok = PayDayLoan.uncache_key(pdl(), 1)
 
     assert [] == PDLTestProcessCache.keys
     assert [] == PDLTestProcessCache.pids
@@ -402,7 +292,7 @@ defmodule ProcessCacheTest do
     key = PDLTestProcessImplementation.key_that_returns_ignore_on_new
     assert {:error, :failed} == PDLTestProcessCache.get_pid(key)
     # should get cleared from the load state cache
-    assert nil == PDL.peek_load_state(PDLTestProcessCache.pdl, key)
+    assert nil == PDL.peek_load_state(pdl(), key)
   end
 
   test "refresh :ignore failures are ignored (should be handled in callback)" do
@@ -419,7 +309,7 @@ defmodule ProcessCacheTest do
     # fail to refresh
     assert {:error, :failed} == PDLTestProcessCache.get_pid(key)
     # should get cleared from the load state cache
-    assert nil == PDL.peek_load_state(PDLTestProcessCache.pdl, key)
+    assert nil == PDL.peek_load_state(pdl(), key)
   end
 
   test "callback_module is a required option for `use PayDayLoan`" do
